Update and rename PAT-B1001.害死人不偿命的(3n + 1)猜想.java to PAT-B1001.害死人不偿命的(3n+1)猜想.java
delete the blank sapce in the title
